add test and local envs & refactor server into separate setup and index files
