+++ src/index.ts
@@ -1,20 +1,6 @@
-import 'reflect-metadata';
-import { AppDataSource } from './data-source';
-import { User } from './entity/User';
+import { setup } from './setup';
+import { config } from 'dotenv';
 
-AppDataSource.initialize()
-  .then(async () => {
-    console.log('Inserting a new user into the database...');
-    const user = new User();
-    user.name = 'Joao Pedro Arroyo';
-    user.password = 'secretpass';
-    user.birthDate = '12/09/2004';
-    user.email = 'joao.arroyo@fake.com';
-    await AppDataSource.manager.save(user);
-    console.log('Saved a new user with id: ' + user.id);
-
-    console.log('Loading users from the database...');
-    const users = await AppDataSource.manager.find(User);
-    console.log('Loaded users: ', users);
-  })
-  .catch((error) => console.log(error));
+config();
+console.info('Initiating setup...');
+setup().catch((error) => console.log(error));
