+++ test.env
@@ -1,0 +1,2 @@
+DATABASE_URL=postgres://test_user:taqtest@localhost:5433/test_db
+PORT=4001
